Factor out service request creation
