+++ src/Factory/JobCoordinator/JobFactory.php
@@ -6,8 +6,6 @@ namespace SmartAssert\ApiClient\Factory\JobCoordinator;
 
 use SmartAssert\ApiClient\Data\JobCoordinator\Job\Components;
 use SmartAssert\ApiClient\Data\JobCoordinator\Job\Job;
-use SmartAssert\ApiClient\Data\JobCoordinator\Job\ServiceRequest;
-use SmartAssert\ApiClient\Data\JobCoordinator\Job\ServiceRequestAttempt;
 use SmartAssert\ApiClient\Exception\IncompleteDataException;
 use SmartAssert\ApiClient\Factory\AbstractFactory;
 
@@ -21,6 +19,7 @@ readonly class JobFactory extends AbstractFactory
         private SerializedSuiteFactory $serializedSuiteFactory,
         private MachineFactory $machineFactory,
         private WorkerJobFactory $workerJobFactory,
+        private ServiceRequestFactory $serviceRequestFactory,
     ) {}
 
     /**
@@ -85,7 +84,7 @@ readonly class JobFactory extends AbstractFactory
         }
 
         try {
-            $serviceRequests = $this->createServiceRequests($serviceRequestDataCollection);
+            $serviceRequests = $this->serviceRequestFactory->createCollection($serviceRequestDataCollection);
         } catch (IncompleteDataException $e) {
             throw new IncompleteDataException($data, 'service_requests.' . $e->missingKey);
         }
@@ -112,73 +111,5 @@ readonly class JobFactory extends AbstractFactory
             new Components($components),
             $serviceRequests,
         );
-    }
-
-    /**
-     * @param array<mixed> $data
-     *
-     * @return ServiceRequest[]
-     *
-     * @throws IncompleteDataException
-     */
-    private function createServiceRequests(array $data): array
-    {
-        $serviceRequests = [];
-
-        foreach ($data as $serviceRequestIndex => $serviceRequestData) {
-            if (is_array($serviceRequestData)) {
-                try {
-                    $serviceRequests[] = $this->createServiceRequest($serviceRequestData);
-                } catch (IncompleteDataException $e) {
-                    throw new IncompleteDataException($data, $serviceRequestIndex . '.' . $e->missingKey);
-                }
-            }
-        }
-
-        return $serviceRequests;
-    }
-
-    /**
-     * @param array<mixed> $data
-     *
-     * @throws IncompleteDataException
-     */
-    private function createServiceRequest(array $data): ServiceRequest
-    {
-        $type = $this->getNonEmptyString($data, 'type');
-
-        $attemptsData = $data['attempts'] ?? null;
-        $attemptsData = is_array($attemptsData) ? $attemptsData : null;
-        if (null === $attemptsData) {
-            throw new IncompleteDataException($data, 'attempts');
-        }
-
-        try {
-            $attempts = $this->createServiceRequestAttempts($attemptsData);
-        } catch (IncompleteDataException $e) {
-            throw new IncompleteDataException($data, 'attempts.' . $e->missingKey);
-        }
-
-        return new ServiceRequest($type, $attempts);
-    }
-
-    /**
-     * @param array<mixed> $data
-     *
-     * @return ServiceRequestAttempt[]
-     *
-     * @throws IncompleteDataException
-     */
-    private function createServiceRequestAttempts(array $data): array
-    {
-        $attempts = [];
-
-        foreach ($data as $attemptData) {
-            if (is_array($attemptData)) {
-                $attempts[] = new ServiceRequestAttempt($this->getNonEmptyString($attemptData, 'state'));
-            }
-        }
-
-        return $attempts;
     }
 }
