+++ tests/Functional/Client/JobCoordinatorClient/AbstractJobCoordinatorClientTestCase.php
@@ -11,6 +11,7 @@ use SmartAssert\ApiClient\Factory\JobCoordinator\MetaStateFactory;
 use SmartAssert\ApiClient\Factory\JobCoordinator\PreparationFactory;
 use SmartAssert\ApiClient\Factory\JobCoordinator\ResultsJobFactory;
 use SmartAssert\ApiClient\Factory\JobCoordinator\SerializedSuiteFactory;
+use SmartAssert\ApiClient\Factory\JobCoordinator\ServiceRequestFactory;
 use SmartAssert\ApiClient\Factory\JobCoordinator\SummaryFactory;
 use SmartAssert\ApiClient\Factory\JobCoordinator\WorkerJobFactory;
 use SmartAssert\ApiClient\JobCoordinatorClient;
@@ -44,6 +45,7 @@ abstract class AbstractJobCoordinatorClientTestCase extends AbstractClientTestCa
                 new SerializedSuiteFactory($metaStateFactory),
                 new MachineFactory($metaStateFactory),
                 new WorkerJobFactory($metaStateFactory),
+                new ServiceRequestFactory(),
             ),
             new SummaryFactory(),
             new HttpHandler(
